requirements.txt
Flask==3.0.3
Werkzeug==3.0.3
Jinja2==3.1.4
sqlalchemy==2.0.35
psycopg[binary]==3.2.1
requests==2.32.3
gunicorn==22.0.0
python-dotenv==1.0.1
